Reorganize files, augment the API
